+++ Assets/Scripts/HP/IDamageable.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IDamageable 
 {
-    public int Hp { get; set; }
+    public float Hp { get; set; }
     void TakeDamage(int damage, LayerMask mask);
     void CheckDeath();
 }
